fix: review fixes for test rank registration, Option A packaging, hardening
Address the review findings on top of the backend-template work:

- tests (P1-1): cast std::abs() result to double in local_max_error so the
  float T deduction failure in test_ppotrf_real builds is fixed.
- tests (P1-2): add an optional MPI rank-count argument to
  add_test_executable; register test_ptran with 6 ranks and
  test_potrf_bottom_right_block / test_potrf_bottom_right with 1 rank so
  ctest launches them with the rank counts they require. Label the manually
  registered test_backend_handle_mpi4 and test_cpu_pgemm_mpi4 as cpu so
  `ctest -L cpu` covers them in CPU-only CI.
- packaging (P2, Option A): drop EXPORT LibDDLATargets / install(EXPORT ...)
  so no LibDDLATargets.cmake is generated or installed; the hand-written
  LibDDLAConfig.cmake remains the single packaging entry point. Install
  headers into CMAKE_INSTALL_INCLUDEDIR instead of a hardcoded
  CMAKE_INSTALL_PREFIX/include (fixes "file INSTALL cannot set permissions
  on /usr/local/include" when installing with --prefix).
- packaging: LibDDLAConfig.cmake now propagates the backend macro
  (DDLA_USE_CUDA / DDLA_USE_HIP / DDLA_USE_CPU, plus CCL/tunnel/debug) and,
  for CPU-enabled builds, the concrete CPU BLAS library path as INTERFACE
  properties on ddla::ddla, so downstream find_package(LibDDLA) consumers
  compile and link out of the box (previously the public headers could not
  compile because local_backend_v was undefined in the consumer TU).
- hardening (P3): MPI_CHECK guards MPI_Abort with MPI_Initialized &&
  !MPI_Finalized; document the CHECK-family collective-hang caveat above
  RUNTIME_CHECK/BLAS_CHECK/SOLVER_CHECK/CCL_CHECK/DERAND_CHECK and the
  tunnel staging-buffer caveats (high-water mark, per-handle
  single-threadedness) in comm_traits.h.

+++ tests/api_grid_test_common.h
@@ -372,7 +372,7 @@ inline double local_max_error(const ddla::DdlaDesc& desc, const std::vector<T>& 
         for(int iloc = 0; iloc < desc.m_loc(); ++iloc){
             const int i = desc.indx_l2g_r(iloc);
             if(i >= desc.m()) continue;
-            err = std::max(err, std::abs(local[iloc + jloc * desc.lld()] - expected(i, j)));
+            err = std::max(err, static_cast<double>(std::abs(local[iloc + jloc * desc.lld()] - expected(i, j))));
         }
     }
     return err;
